Refactored CreateCaseCallbackService
