+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CreateCaseCallbackService.java
@@ -1,79 +1,41 @@
 package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
 
 import com.google.common.base.Strings;
-import com.google.common.collect.ImmutableMap;
 import io.vavr.collection.Seq;
 import io.vavr.control.Try;
 import io.vavr.control.Validation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.InvalidCaseDataException;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.TransformationClient;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.request.ExceptionRecord;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.CaseCreationDetails;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.SuccessfulTransformationResponse;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.ServiceConfigItem;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.model.in.CcdCallbackRequest;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.CreateCaseValidator;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.ProcessResult;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.YesNoFieldValues;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.config.ServiceConfigProvider;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.payments.IPaymentsPublisher;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.payments.model.UpdatePaymentsCommand;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
-import static java.util.stream.Collectors.joining;
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.hasServiceNameInCaseTypeId;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.EventIdValidator.isCreateNewCaseEvent;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.CASE_REFERENCE;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.DISPLAY_WARNINGS;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.OCR_DATA_VALIDATION_WARNINGS;
 
 @Service
 public class CreateCaseCallbackService {
 
     private static final Logger log = LoggerFactory.getLogger(CreateCaseCallbackService.class);
 
-    private static final String EXCEPTION_RECORD_REFERENCE = "bulkScanCaseReference";
-
     private final CreateCaseValidator validator;
     private final ServiceConfigProvider serviceConfigProvider;
-    private final TransformationClient transformationClient;
-    private final AuthTokenGenerator s2sTokenGenerator;
-    private final IPaymentsPublisher paymentsPublisher;
-    private final CoreCaseDataApi coreCaseDataApi;
-    private final CcdApi ccdApi;
+    private final CcdCaseCreationManager ccdCaseCreationManager;
 
     public CreateCaseCallbackService(
         CreateCaseValidator validator,
         ServiceConfigProvider serviceConfigProvider,
-        TransformationClient transformationClient,
-        AuthTokenGenerator s2sTokenGenerator,
-        IPaymentsPublisher paymentsPublisher,
-        CoreCaseDataApi coreCaseDataApi,
-        CcdApi ccdApi
+        CcdCaseCreationManager ccdCaseCreationManager
     ) {
         this.validator = validator;
         this.serviceConfigProvider = serviceConfigProvider;
-        this.transformationClient = transformationClient;
-        this.s2sTokenGenerator = s2sTokenGenerator;
-        this.paymentsPublisher = paymentsPublisher;
-        this.coreCaseDataApi = coreCaseDataApi;
-        this.ccdApi = ccdApi;
+        this.ccdCaseCreationManager = ccdCaseCreationManager;
     }
 
     /**
@@ -96,7 +58,7 @@ public class CreateCaseCallbackService {
 
         ProcessResult result = validator
             .getValidation(exceptionRecordData)
-            .map(exceptionRecord -> tryCreateNewCase(
+            .map(exceptionRecord -> ccdCaseCreationManager.tryCreateNewCase(
                 exceptionRecord,
                 serviceConfigItem,
                 request.isIgnoreWarnings(),
@@ -141,217 +103,5 @@ public class CreateCaseCallbackService {
         )
             .filter(item -> !Strings.isNullOrEmpty(item.getTransformationUrl()))
             .getOrElse(Validation.invalid("Transformation URL is not configured"));
-    }
-
-    private ProcessResult tryCreateNewCase(
-        ExceptionRecord exceptionRecord,
-        ServiceConfigItem configItem,
-        boolean ignoreWarnings,
-        String idamToken,
-        String userId,
-        CaseDetails exceptionRecordData
-    ) {
-        List<Long> ids = ccdApi.getCaseRefsByBulkScanCaseReference(exceptionRecord.id, configItem.getService());
-        if (ids.isEmpty()) {
-            return createNewCase(
-                exceptionRecord,
-                configItem,
-                ignoreWarnings,
-                idamToken,
-                userId,
-                exceptionRecordData
-            );
-        } else if (ids.size() == 1) {
-            return new ProcessResult(
-                ImmutableMap.<String, Object>builder()
-                    .put(CASE_REFERENCE, Long.toString(ids.get(0)))
-                    .put(DISPLAY_WARNINGS, YesNoFieldValues.NO)
-                    .put(OCR_DATA_VALIDATION_WARNINGS, emptyList())
-                    .build()
-            );
-        } else {
-            return new ProcessResult(
-                emptyList(),
-                singletonList(
-                    String.format(
-                        "Multiple cases (%s) found for the given bulk scan case reference: %s",
-                        ids.stream().map(String::valueOf).collect(joining(", ")),
-                        exceptionRecord.id
-                    )
-                )
-            );
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private ProcessResult createNewCase(
-        ExceptionRecord exceptionRecord,
-        ServiceConfigItem configItem,
-        boolean ignoreWarnings,
-        String idamToken,
-        String userId,
-        CaseDetails exceptionRecordData
-    ) {
-        log.info(
-            "Start creating new case for {} from exception record {}",
-            configItem.getService(),
-            exceptionRecord.id
-        );
-
-        try {
-            String s2sToken = s2sTokenGenerator.generate();
-
-            SuccessfulTransformationResponse transformationResponse = transformationClient.transformExceptionRecord(
-                configItem.getTransformationUrl(),
-                exceptionRecord,
-                s2sToken
-            );
-
-            if (!ignoreWarnings && !transformationResponse.warnings.isEmpty()) {
-                // do not log warnings
-                return new ProcessResult(transformationResponse.warnings, emptyList());
-            }
-
-            log.info(
-                "Successfully transformed exception record for {} from exception record {}",
-                configItem.getService(),
-                exceptionRecord.id
-            );
-
-            checkBulkScanReferenceIsSet(
-                (Map<String, ?>) transformationResponse.caseCreationDetails.caseData,
-                exceptionRecord.id
-            );
-
-            long newCaseId = createNewCaseInCcd(
-                idamToken,
-                s2sToken,
-                userId,
-                exceptionRecord,
-                transformationResponse.caseCreationDetails
-            );
-
-            log.info(
-                "Successfully created new case for {} with case ID {} from exception record {}",
-                configItem.getService(),
-                newCaseId,
-                exceptionRecord.id
-            );
-
-            handlePayments(exceptionRecordData, newCaseId);
-
-            return new ProcessResult(
-                ImmutableMap.<String, Object>builder()
-                    .put(CASE_REFERENCE, Long.toString(newCaseId))
-                    .put(DISPLAY_WARNINGS, YesNoFieldValues.NO)
-                    .put(OCR_DATA_VALIDATION_WARNINGS, emptyList())
-                    .build()
-            );
-        } catch (InvalidCaseDataException exception) {
-            if (BAD_REQUEST.equals(exception.getStatus())) {
-                throw exception;
-            } else {
-                return new ProcessResult(exception.getResponse().warnings, exception.getResponse().errors);
-            }
-        } catch (Exception exception) {
-            log.error(
-                "Failed to create exception for service {} and exception record {}",
-                configItem.getService(),
-                exceptionRecord.id,
-                exception
-            );
-
-            return new ProcessResult(emptyList(), singletonList("Internal error. " + exception.getMessage()));
-        }
-    }
-
-    private void handlePayments(CaseDetails exceptionRecord, long newCaseId) {
-
-        boolean containsPayments =
-            Objects.equals(
-                exceptionRecord.getData().get(ExceptionRecordFields.CONTAINS_PAYMENTS),
-                YesNoFieldValues.YES
-            );
-
-        if (containsPayments) {
-
-            String envelopeId = exceptionRecord.getData().get(ExceptionRecordFields.ENVELOPE_ID).toString();
-            String jurisdiction = exceptionRecord.getData().get(ExceptionRecordFields.PO_BOX_JURISDICTION).toString();
-
-            log.info(
-                "Sending payment update message. ER id: {}",
-                exceptionRecord.getId()
-            );
-
-            paymentsPublisher.send(
-                new UpdatePaymentsCommand(
-                    Long.toString(exceptionRecord.getId()),
-                    Long.toString(newCaseId),
-                    envelopeId,
-                    jurisdiction
-                )
-            );
-        } else {
-            log.info(
-                "Exception record has no payments, not sending update command. ER id: {}",
-                exceptionRecord.getId()
-            );
-        }
-    }
-
-    private long createNewCaseInCcd(
-        String idamToken,
-        String s2sToken,
-        String userId,
-        ExceptionRecord exceptionRecord,
-        CaseCreationDetails caseCreationDetails
-    ) {
-        StartEventResponse eventResponse = coreCaseDataApi.startForCaseworker(
-            idamToken,
-            s2sToken,
-            userId,
-            exceptionRecord.poBoxJurisdiction,
-            caseCreationDetails.caseTypeId,
-            // when onboarding remind services to not configure about to submit callback for this event
-            caseCreationDetails.eventId
-        );
-
-        return coreCaseDataApi.submitForCaseworker(
-            idamToken,
-            s2sToken,
-            userId,
-            exceptionRecord.poBoxJurisdiction,
-            caseCreationDetails.caseTypeId,
-            true,
-            CaseDataContent
-                .builder()
-                .caseReference(exceptionRecord.id)
-                .data(caseCreationDetails.caseData)
-                .event(Event
-                    .builder()
-                    .id(eventResponse.getEventId())
-                    .summary("Case created")
-                    .description("Case created from exception record ref " + exceptionRecord.id)
-                    .build()
-                )
-                .eventToken(eventResponse.getToken())
-                .build()
-        ).getId();
-    }
-
-    private void checkBulkScanReferenceIsSet(Map<String, ?> caseData, String exceptionRecordId) {
-        if (!caseData.containsKey(EXCEPTION_RECORD_REFERENCE)) {
-            log.error(
-                "Transformation did not set '{}' with exception record id {}",
-                EXCEPTION_RECORD_REFERENCE,
-                exceptionRecordId
-            );
-        } else if (!caseData.get(EXCEPTION_RECORD_REFERENCE).equals(exceptionRecordId)) {
-            log.error(
-                "Transformation did not set exception record reference correctly. Actual: {}, expected: {}",
-                caseData.get(EXCEPTION_RECORD_REFERENCE),
-                exceptionRecordId
-            );
-        }
     }
 }
